+++ src/Classes/Latte/EngineBuilder.php
@@ -2,13 +2,15 @@
 
 declare(strict_types=1);
 
-namespace CloudBase\LatteHelper;
+namespace CloudBase\LatteHelper\Classes\Latte;
 
 use Latte\Engine;
 use Latte\Loaders\FileLoader;
 
 class EngineBuilder
 {
+    private const string EXTENSIONS_CONFIG_PATH = '/config/latte.php';
+
     public static function getEngine(string $templateDir): Engine
     {
         if (!str_starts_with($templateDir, '/')) {
@@ -21,6 +23,10 @@ class EngineBuilder
         $engine->setTempDirectory(EngineBuilder::getProjectDirectory() . '/var/cache/latte');
         $engine->setLoader($fileLoader);
 
+        foreach (EngineBuilder::loadConfig() as $class => $args) {
+            $engine->addExtension(new $class(...$args));
+        }
+
         return $engine;
     }
 
@@ -31,5 +37,24 @@ class EngineBuilder
         }
 
         return dirname(__DIR__);
+    }
+
+    /**
+     * @return array<string, array<int, mixed>>
+     */
+    private static function loadConfig(): array
+    {
+        $config = [];
+        $configPath = sprintf(
+            '%s/%s',
+            EngineBuilder::getProjectDirectory(),
+            EngineBuilder::EXTENSIONS_CONFIG_PATH
+        );
+
+        if (file_exists($configPath)) {
+            $config = require $configPath;
+        }
+
+        return is_array($config) ? $config : [];
     }
 }