+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.0] - 2025-10-21
+### Added
+- Added `LatteAwareApplication` to provide `$app` to templates.
+- Provides: `getUser`, `getSession`, `getRequestStack`, `getFlashes` methods.
+- Added the ability to register custom Latte extensions via a `config/latte.php` file in your Symfony project.
+
+### Changed
+- Changed `AbstractLatteController::globalData` method to always pass an instance of `LatteAwareApplication`. Ensure to
+  merge `parent::globalData()` into your own data if you require `$app` overriding the method in your controllers.
+
 ## [0.2.0] - 2025-10-20
 ### Fixed
 - `AbstractLatteController` now extends `AbstractController`.